+++ src/main/resources/assets/ae2fc/lang/en_US.lang
@@ -30,8 +30,8 @@ item.fluid_storage.256.name=%s256k%s ME Fluid Storage Cell
 item.fluid_storage.1024.name=%s1024k%s ME Fluid Storage Cell
 item.fluid_storage.4096.name=%s4096k%s ME Fluid Storage Cell
 item.fluid_storage.16384.name=%s16384k%s ME Fluid Storage Cell
-item.fluid_storage.131072.name=%s131072k%s ME Fluid Quantum Storage Cell
-item.fluid_storage.554160890921312.name=%s554P%s ME Fluid Digital Singularity Storage Cell
+item.fluid_storage.quantum.name=ME Fluid Quantum Storage Cell
+item.fluid_storage.singularity.name=ME Fluid Digital Singularity Storage Cell
 item.fluid_part.0.name=%s1k%s ME Fluid Storage Component
 item.fluid_part.1.name=%s4k%s ME Fluid Storage Component
 item.fluid_part.2.name=%s16k%s ME Fluid Storage Component
