+++ src/main/resources/assets/ae2fc/lang/zh_CN.lang
@@ -30,8 +30,8 @@ item.fluid_storage.256.name=%s256k%s-ME流体存储元件
 item.fluid_storage.1024.name=%s1024k%s-ME流体存储元件
 item.fluid_storage.4096.name=%s4096k%s-ME流体存储元件
 item.fluid_storage.16384.name=%s16384k%s-ME流体存储元件
-item.fluid_storage.131072.name=%s131072k%s ME流体量子存储元件
-item.fluid_storage.554160890921312.name=%s554P%s MEl流体数字奇点元件
+item.fluid_storage.quantum.name=ME流体量子存储元件
+item.fluid_storage.singularity.name=ME流体数字奇点元件
 item.fluid_part.0.name=%s1k%s-ME流体存储组件
 item.fluid_part.1.name=%s4k%s-ME流体存储组件
 item.fluid_part.2.name=%s16k%s-ME流体存储组件
