+++ src/main/resources/assets/ae2fc/textures/items/fluid_storage.singularity.png
